pyproject.toml
[project]
name = "great-minds"
version = "0.1.0"
description = "LLM-powered research knowledge base framework"
requires-python = ">=3.14"
dependencies = [
    "absurd-sdk>=0.3.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "boto3>=1.42.88",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "markitdown>=0.1.5",
    "openai>=2.0.0",
    "pgvector>=0.4.2",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.0",
    "python-multipart>=0.0.24",
    "resend>=2.27.0",
    "ruamel.yaml>=0.19.0",
    "scikit-learn>=1.5.0",
    "sqlalchemy[asyncio]>=2.0.49",
    "uuid6>=2024.7.10",
    "uvicorn>=0.44.0",
    "rich>=14.3.3",
    "typer>=0.24.1",
    "numpy>=2.4.4",
]

[project.scripts]
great-minds = "great_minds.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
ignore = ["E712"]

[dependency-groups]
dev = [
    "moto[s3]>=5.1.22",
    "ruff>=0.15.9",
    "ty>=0.0.29",
]
